fix: edit YouTubeParser
